+++ book/examples/smoothing/smoothing.py
