+++ book/examples/smoothing/smoothing_types.py
